Fix logic errors for contact syncing
Fixes bugs:
- When there are multiple types of contacts to be modified (update,
add, remove), the code only used to handle one type (if,else if..).
Now all types are handled during one pass.
- There is some bug with some displayNames being deleted every
time, so I moved the removal of users to the very beginning, and
the code checks for contacts again at the end for contacts to be
added. This way erroneously removed users which are present in the
GAL will be re-added.